Update Fast-CDR thirdparty
If this commit is applied then plotjuggler will be able to compile and
run with Fast-CDR version 2.x.x from eProsima.
It can compile from:
- local thirdparty folder included within the repo
- install of fastcdr within the machine (usually /usr/local)

to do so:
- Udpated files, API and namespace names

Done it because I need to use plotJuggler with latest lib from eProsima
and I would like this to be available for everyone.

Issue:
[https://github.com/facontidavide/PlotJuggler/issues/919]